statuspage-bot/requirements.txt
slack_bolt==1.18.0
python-dotenv==1.0.0
requests==2.30.0